ログファイルのパスを修正+tmpファイルがgitに残らないように修正

+++ src/http11.cpp
@@ -313,7 +313,7 @@ int DoPath11(Socket* sClient, int iMethod, char* szPath, char* szSearch,
   // Send full entity.
   sprintf(szBuf, "Content-Type: %s\r\n", eExtMap[iType].szType);
   sClient->Send(szBuf);
-  sprintf(szBuf, "Content-Length: %ld\r\n", sBuf.st_size);
+  sprintf(szBuf, "Content-Length: %ld\r\n", (long)sBuf.st_size);
   sClient->Send(szBuf);
   sClient->Send("\r\n");
 
@@ -526,7 +526,7 @@ int DoExec11(Socket* sClient, int iMethod, char* szPath, char* szSearch,
   }
   ifIn.close();
   iCount += 2; // The last CRLF isn't counted within the loop.
-  sprintf(szBuf, "Content-Length: %ld\r\n\r\n", sBuf.st_size - iCount);
+  sprintf(szBuf, "Content-Length: %ld\r\n\r\n", (long)sBuf.st_size - iCount);
   sClient->Send(szBuf);
 
   if (iMethod != HEAD) // Only send the entity if not HEAD.
@@ -943,7 +943,7 @@ int SendByteRange(Socket* sClient, Headers* hInfo, char* szPath,
       sprintf(szBuf, "Content-Type: %s\r\n", eExtMap[iType].szType);
       sClient->Send(szBuf); // Now content-type.
       sprintf(szBuf, "Content-Range: bytes %d-%d/%ld\r\n\r\n",
-          hInfo->rRanges[i].iStart, hInfo->rRanges[i].iEnd, sBuf->st_size);
+          hInfo->rRanges[i].iStart, hInfo->rRanges[i].iEnd, (long)sBuf->st_size);
       sClient->Send(szBuf); // Now content-range.
 
       ifIn.seekg(hInfo->rRanges[i].iStart, std::ios::beg);
